refactor(UserDataStore): extract WalkingSpeed types to dedicated file
`UserDataStore.swift` was over the SwiftLint 1100-line error threshold.
Move `WalkingSpeedSource` and the `WalkingSpeed` namespace into a new
co-conceptual file under `Models/UserData/`.
